ROJO DE NUEVO-Se modifica el tipo de dato del atributo pero no la asignacion en el metodo.

+++ EjercicioTDD/cuentaBancaria/CuentaBancaria.java
@@ -1,7 +1,7 @@
 package cuentaBancaria;
 
 public class CuentaBancaria {
-	private int monto;
+	private double monto;
 	private long numCuenta;
 	
 	public void realizarTransferencia () {
